refactor: break circular dependency between job-tools.ts and server.ts
Move toToolError into dedicated tool-error.ts and JobTools type into
job-tools.ts so the extracted module no longer imports from server.ts.
Dependency direction is now server.ts → job-tools.ts → tool-error.ts
with no cycles.
