+++ apps/server/src/shared/mcp/job-tools.ts
@@ -1,8 +1,85 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import * as z from "zod/v4";
 
-import type { JobTools } from "./server.js";
-import { toToolError } from "./server.js";
+import { toToolError } from "./tool-error.js";
+
+export type JobTools = {
+  complete: (
+    agentId: string,
+    report: unknown
+  ) => Promise<{ runId: string; status: string }>;
+  failed: (
+    agentId: string,
+    report: unknown
+  ) => Promise<{ runId: string; status: string }>;
+  needsInput: (
+    agentId: string,
+    question: string
+  ) => Promise<{ runId: string; status: string }>;
+  log: (
+    agentId: string,
+    input: {
+      task: string;
+      message: string;
+      level: "debug" | "info" | "warn" | "error";
+    }
+  ) => Promise<{ runId: string; status: string }>;
+  listAgents: () => Promise<
+    Array<{ id: string; name: string; status: string; cwd: string }>
+  >;
+  listRecentPersonaReviews: (sinceDays: number) => Promise<
+    Array<{
+      id: number;
+      agentId: string;
+      parentAgentId: string;
+      persona: string;
+      status: string;
+      message: string | null;
+      verdict: string | null;
+      summary: string | null;
+      filesReviewed: string[] | null;
+      createdAt: string;
+      updatedAt: string;
+    }>
+  >;
+  listRecentFeedback: (sinceDays: number) => Promise<
+    Array<{
+      id: number;
+      agentId: string;
+      persona: string;
+      severity: string;
+      filePath: string | null;
+      lineNumber: number | null;
+      description: string;
+      suggestion: string | null;
+      mediaRef: string | null;
+      status: string;
+      createdAt: string;
+    }>
+  >;
+  getActivitySummary: (params: {
+    start: Date;
+    end: Date;
+    project?: string;
+  }) => Promise<Record<string, unknown>>;
+  getAgentHistory: (params: {
+    start: Date;
+    end: Date;
+    project?: string;
+    limit: number;
+    offset: number;
+    includeEvents: boolean;
+    includeFeedback: boolean;
+    includeReviews: boolean;
+    includeChildren: boolean;
+  }) => Promise<Record<string, unknown>>;
+  getFeedbackSummary: (params: {
+    start: Date;
+    end: Date;
+    project?: string;
+    groupBy: "persona" | "severity" | "directory";
+  }) => Promise<Record<string, unknown>>;
+};
 
 export function registerJobTools(
   server: McpServer,
